requirements.txt
torch
numpy
tqdm
POT
Pillow
torchvision
opencv-python
